+++ src/main/kotlin/icu/windea/sqlfox/SqlFoxApplication.kt
@@ -1,5 +1,6 @@
 package icu.windea.sqlfox
 
+import com.alibaba.druid.DbType
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
 
@@ -9,3 +10,5 @@ class SqlFoxApplication
 fun main(args: Array<String>) {
     runApplication<SqlFoxApplication>(*args)
 }
+
+val defaultDbType = DbType.mysql