+++ src/main/kotlin/icu/windea/sqlfox/component/HelloWorldComponent.kt
@@ -4,10 +4,11 @@ import org.springframework.shell.standard.*
 
 @ShellComponent
 class HelloWorldComponent {
-    @ShellMethod("Test command")
+    @ShellMethod("Test command.")
     fun hello(
         @ShellOption(defaultValue = "world", help = "Say hello to who?") who: String
     ): String {
         return "hello $who"
     }
 }
+