+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 ## 1.0
 
 * 初始化仓库
+* 部分实现命令`format`，实现选项`lower-case` `upper-case`